requirements.txt
flask
requests
python-dotenv
sentence-transformers
faiss-cpu
transformers
torch